Refactor SharePoint site fetching logic by removing unused auth type and simplifying error handling
